Cast to `size_t` to avoid int overflow in `ht_undo_impl`
`width` is `int32_t` and `bytes_per_element` is `int16_t`, so the raw
product can overflow 32 bits before the cast to `size_t`; cast each
term to `size_t` before the multiplication.

+++ src/lib/OpenEXRCore/internal_ht.cpp
@@ -185,8 +185,8 @@ ht_undo_impl (
 
         size_t computedoffset = 0;
         for (int i = 0; i < file_i; ++i)
-            computedoffset += decode->channels[i].width *
-                              decode->channels[i].bytes_per_element;
+            computedoffset += (size_t) decode->channels[i].width *
+                              (size_t) decode->channels[i].bytes_per_element;
         cs_to_file_ch[cs_i].raster_line_offset = computedoffset;
     }
 
@@ -212,8 +212,8 @@ ht_undo_impl (
     bool is_planar  = false;
     for (int16_t c = 0; c < decode->channel_count; c++)
     {
-        bpl +=
-            decode->channels[c].bytes_per_element * decode->channels[c].width;
+        bpl += (int64_t) decode->channels[c].bytes_per_element *
+               (int64_t) decode->channels[c].width;
         if (decode->channels[c].x_samples > 1 ||
             decode->channels[c].y_samples > 1)
         { is_planar = true; }
@@ -278,8 +278,8 @@ ht_undo_impl (
                         }
                     }
 
-                    line_pixels += decode->channels[line_c].bytes_per_element *
-                                   decode->channels[line_c].width;
+                    line_pixels += (size_t) decode->channels[line_c].bytes_per_element *
+                                   (size_t) decode->channels[line_c].width;
                 }
             }
         }
